requirements.txt
numpy>=1.23.5
opencv-python-headless==4.10.0.84
pandas==2.2.0
Pillow==10.3.0
torch==2.3.1				
torchvision==0.18.1			
efficientnet_pytorch==0.7.1
tqdm==4.66.4
scipy==1.14.0
scikit-image==0.24.0
h5py==3.11.0
matplotlib==3.9.0
timm==1.0.15
numba==0.61.0
Bottleneck==1.4.2
SimpleITK==4.2.1
